Allow a custom AppDelegate to be set on macOS.
The official godot Application delegate on macOS expects the display server to be macOS and crashes when delegate events happen. This replaces the delegate with one that handles events Godot requires manually.

It also allows custom delegates to be set, while maintaining the event handling that Godot expects

The delegate is set *after* the application starts so it will miss some events like (applicationDidFinishLaunching).